requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.2.0
email-validator>=2.0.0
python-dotenv>=1.0.0
flask-limiter>=3.5.0
gunicorn>=22.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
